feat: create an interface for the API Caller
This provide an on-disk implementation of the API Caller depending on
the use. Currently it helps to develop locally while keeping the API
requests to a minimum. It's also useful when working offline.

Currently, only the basic "get me all notifications" request is handled,
and further implementations will be done when necessary.