feat(build): simplify build/install process by pulling debug info
Starting in 1.18, git information is stored in the binary, it's possible
to pull it at runtime. Using the runtime/debug package, we can extract
the `commit` and `date` information.

The `version` (renamed `tag`) still comes from the ldflags. This allows
for simpler build/install process.

- https://pkg.go.dev/runtime/debug@go1.22.5#ReadBuildInfo
- https://icinga.com/blog/2022/05/25/embedding-git-commit-information-in-go-binaries/

Thanks @natefinch, @chrisgavin, @cogswell for the help.

Fix #95