fix(config): warn about tilde expansion not happening in Go
The `~` expansion doesn't happen in `gh-not`, and should happen in the
shell anyway. At least in bash & zsh, the `~` doesn't expand if it's not preceded by a
space character.

In opposition, $HOME is always expanded before being sent to the
program.

E.g.

```shell
echo $*
```

```shell
a.sh --a=~/test/a      => --a=~/test/a
a.sh --a ~/test/a      => --a /home/user/test/a
a.sh --a $HOME/test/a  => --a /home/user/test/a
a.sh --a=$HOME/test/a  => --a /home/user/test/a
```

This adds a warning if the config path is found to start with a `~`.

It also adds a log of the flags as they are received by `gh-not` for
additional debugging.

Fix #246