fix: strip potential newlines from subject's title
It was found during an exploration with @ekroon that it's possible for
PR/Issue titles to contain newline characters.

E.g.

```shell
$ gh api /repos/X/Y/issues -X POST -F title="this is a test
"

$ gh api /repos/X/Y/issues/Z | jq '.title'
"this is a test\n"
```

This would effectively break the table rendering, as it would create
extra lines to be saved. The simplest solution is to strip those
characters from the title before rendering it to the terminal.

Fix #288